Start on v2 of role mgr
