+++ routing.ts
@@ -1,23 +1,13 @@
-/*
-services from jacdac-v0
-
-debugging services?
-name service - need to re-implement
-identification service - led blinking
-
-*/
-
 namespace jacdac {
-    const devNameSettingPrefix = "#jddev:"
     // common logging level for jacdac services
     export let consolePriority = ConsolePriority.Debug;
 
     let _hostServices: Host[]
     let _unattachedClients: Client[]
-    let _allClients: Client[]
+    export let _allClients: Client[]
     let _myDevice: Device
     //% whenUsed
-    let _devices: Device[] = []
+    export let _devices: Device[] = []
     //% whenUsed
     let _announceCallbacks: (() => void)[] = [];
     let _newDeviceCallbacks: (() => void)[];
@@ -328,7 +318,7 @@ namespace jacdac {
         _attach(dev: Device, serviceNum: number) {
             if (this.device) throw "Invalid attach"
             if (!this.broadcast) {
-                if (this.requiredDeviceName && this.requiredDeviceName != dev.name && this.requiredDeviceName != dev.deviceId)
+                if (!dev.matchesNameAt(this.requiredDeviceName, serviceNum))
                     return false // don't attach
                 this.device = dev
                 this.serviceNumber = serviceNum
@@ -453,7 +443,6 @@ namespace jacdac {
         services: Buffer
         lastSeen: number
         clients: Client[] = []
-        private _name: string
         private _shortId: string
         private queries: RegQuery[]
 
@@ -465,13 +454,6 @@ namespace jacdac {
             return this.clients != null
         }
 
-        get name() {
-            // TODO measure if caching is worth it
-            if (this._name === undefined)
-                this._name = settings.readString(devNameSettingPrefix + this.deviceId) || null
-            return this._name
-        }
-
         get shortId() {
             // TODO measure if caching is worth it
             if (!this._shortId)
@@ -480,7 +462,21 @@ namespace jacdac {
         }
 
         toString() {
-            return this.shortId + (this.name ? ` (${this.name})` : ``)
+            return this.shortId
+        }
+
+        matchesNameAt(name: string, serviceNum: number) {
+            if (!name)
+                return true
+
+            if (name == this.deviceId)
+                return true
+            if (name == this.deviceId + ":" + serviceNum)
+                return true
+
+            const id = name + ":" + serviceNum
+            const name2 = settings.readString(id)
+            return name == name2
         }
 
         private lookupQuery(reg: number) {
@@ -554,8 +550,6 @@ namespace jacdac {
         }
 
         static clearNameCache() {
-            for (let d of _devices)
-                d._name = undefined
             clearAttachCache()
         }
 
@@ -683,7 +677,7 @@ namespace jacdac {
                 continue // will re-attach
             }
             const newClass = dev.services.getNumber(NumberFormat.UInt32LE, c.serviceNumber << 2)
-            if (newClass == c.serviceClass && (!c.requiredDeviceName || c.requiredDeviceName == dev.name)) {
+            if (newClass == c.serviceClass && dev.matchesNameAt(c.requiredDeviceName, c.serviceNumber)) {
                 newClients.push(c)
                 occupied[c.serviceNumber] = 1
             } else {
@@ -911,215 +905,5 @@ namespace jacdac {
             roleManagerHost.start();
         // and we're done
         log("jacdac started");
-    }
-
-    export function autoBind() {
-        function log(msg: string) {
-            control.dmesg("autobind: " + msg)
-        }
-
-        function pending() {
-            return _allClients.filter(c => !!c.requiredDeviceName && !c.isConnected())
-        }
-
-        pauseUntil(() => pending().length == 0, 1000)
-
-        const plen = pending().length
-        log(`pending: ${plen}`)
-        if (plen == 0) return
-
-        pause(1000) // wait for everyone to enumerate
-
-        const requested: RemoteRequestedDevice[] = []
-
-        for (const client of _allClients) {
-            if (client.requiredDeviceName) {
-                const r = addRequested(requested, client.requiredDeviceName, client.serviceClass, null)
-                r.boundTo = client.device
-            }
-        }
-
-        if (!requested.length)
-            return
-
-        function nameFree(d: Device) {
-            return !d.name || requested.every(r => r.boundTo != d)
-        }
-
-        requested.sort((a, b) => a.name.compare(b.name))
-
-        let numSel = 0
-        recomputeCandidates(requested)
-        for (const r of requested) {
-            if (r.boundTo)
-                continue
-            const cand = r.candidates.filter(nameFree)
-            log(`name: ${r.name}, ${cand.length} candidate(s)`)
-            if (cand.length > 0) {
-                // take ones without existing names first
-                cand.sort((a, b) => (a.name || "").compare(b.name || "") || a.deviceId.compare(b.deviceId))
-                log(`setting to ${cand[0].toString()}`)
-                r.select(cand[0])
-                numSel++
-            }
-        }
-    }
-
-    function clearAllNames() {
-        settings.list(devNameSettingPrefix).forEach(settings.remove)
-    }
-
-    function setDevName(id: string, name: string) {
-        const devid = devNameSettingPrefix + id
-        if (name.length == 0)
-            settings.remove(devid)
-        else
-            settings.writeString(devid, name)
-        Device.clearNameCache()
-    }
-
-    export class RoleManagerHost extends Host {
-        constructor() {
-            super("rolemgr", SRV_ROLE_MANAGER)
-        }
-
-        public handlePacket(packet: JDPacket) {
-            switch (packet.service_command) {
-                case RoleManagerCmd.GetRole:
-                    if (packet.data.length == 8) {
-                        let name = settings.readBuffer(devNameSettingPrefix + packet.data.toHex())
-                        if (!name) name = Buffer.create(0)
-                        this.sendReport(JDPacket.from(RoleManagerCmd.GetRole, packet.data.concat(name)))
-                    }
-                    break
-                case RoleManagerCmd.SetRole:
-                    if (packet.data.length >= 8) {
-                        setDevName(packet.data.slice(0, 8).toHex(), packet.data.slice(8).toString())
-                        this.sendChangeEvent();
-                    }
-                    break
-                case RoleManagerCmd.ListStoredRoles:
-                    OutPipe.respondForEach(packet, settings.list(devNameSettingPrefix), k =>
-                        Buffer.fromHex(k.slice(devNameSettingPrefix.length))
-                            .concat(settings.readBuffer(k)))
-                    break
-                case RoleManagerCmd.ListRequiredRoles:
-                    const namedClients = _allClients.filter(c => !!c.requiredDeviceName)
-                    OutPipe.respondForEach(packet, namedClients, packName)
-                    break
-                case RoleManagerCmd.ClearAllRoles:
-                    clearAllNames()
-                    this.sendChangeEvent();
-                    break
-            }
-
-            function packName(c: Client) {
-                const devid = c.device ? Buffer.fromHex(c.device.deviceId) : Buffer.create(8)
-                return jdpack("b[8] u32 s", [devid, c.serviceClass, c.requiredDeviceName])
-            }
-        }
-    }
-
-    //% fixedInstance whenUsed block="role manager"
-    export const roleManagerHost = new RoleManagerHost()
-
-    export class RemoteRequestedDevice {
-        services: number[] = [];
-        boundTo: Device;
-        candidates: Device[] = [];
-
-        constructor(
-            public parent: RoleManagerClient,
-            public name: string
-        ) { }
-
-        isCandidate(ldev: Device) {
-            return this.services.every(s => ldev.hasService(s))
-        }
-
-        select(dev: Device) {
-            if (dev == this.boundTo)
-                return
-            if (this.parent == null) {
-                setDevName(dev.deviceId, this.name)
-            } else {
-                if (this.boundTo)
-                    this.parent.setName(this.boundTo, "")
-                this.parent.setName(dev, this.name)
-            }
-            this.boundTo = dev
-        }
-    }
-
-    function recomputeCandidates(remotes: RemoteRequestedDevice[]) {
-        const localDevs = devices()
-        for (let dev of remotes)
-            dev.candidates = localDevs.filter(ldev => dev.isCandidate(ldev))
-    }
-
-    function addRequested(devs: RemoteRequestedDevice[], name: string, service_class: number,
-        parent: RoleManagerClient) {
-        let r = devs.find(d => d.name == name)
-        if (!r)
-            devs.push(r = new RemoteRequestedDevice(parent, name))
-        r.services.push(service_class)
-        return r
-    }
-
-
-    export class RoleManagerClient extends Client {
-        public remoteRequestedDevices: RemoteRequestedDevice[] = []
-
-        constructor(requiredDevice: string = null) {
-            super("rolemgrc", SRV_ROLE_MANAGER, requiredDevice)
-
-            onNewDevice(() => {
-                recomputeCandidates(this.remoteRequestedDevices)
-            })
-
-            onAnnounce(() => {
-                if (this.isConnected())
-                    control.runInParallel(() => this.scanCore())
-            })
-        }
-
-        private scanCore() {
-            const inp = new InPipe()
-            this.sendCommand(inp.openCommand(RoleManagerCmd.ListRequiredRoles))
-
-            const localDevs = devices()
-            const devs: RemoteRequestedDevice[] = []
-
-            inp.readList(buf => {
-                const [devidbuf, service_class, name] = jdunpack<[Buffer, number, string]>(buf, "b[8] u32 s")
-                const devid = devidbuf.toHex();
-                const r = addRequested(devs, name, service_class, this)
-                const dev = localDevs.find(d => d.deviceId == devid)
-                if (dev)
-                    r.boundTo = dev
-            })
-
-            devs.sort((a, b) => a.name.compare(b.name))
-
-            this.remoteRequestedDevices = devs
-            recomputeCandidates(this.remoteRequestedDevices)
-        }
-
-        scan() {
-            pauseUntil(() => this.isConnected())
-            this.scanCore()
-        }
-
-        clearNames() {
-            this.sendCommandWithAck(JDPacket.onlyHeader(RoleManagerCmd.ClearAllRoles))
-        }
-
-        setName(dev: Device, name: string) {
-            this.sendCommandWithAck(JDPacket.from(RoleManagerCmd.SetRole,
-                Buffer.fromHex(dev.deviceId).concat(Buffer.fromUTF8(name))))
-        }
-
-        handlePacket(pkt: JDPacket) {
-        }
     }
 }
